corrected import reference cycle issues
