+++ flowsa/data_source_scripts/BLS_QCEW.py
@@ -15,7 +15,8 @@ import pandas as pd
 import numpy as np
 import io
 import zipfile
-from flowsa.common import log, get_all_state_FIPS_2, US_FIPS
+from flowsa.common import log, get_all_state_FIPS_2, US_FIPS, fba_default_grouping_fields, \
+    fba_mapped_default_grouping_fields
 from flowsa.flowbyfunctions import assign_fips_location_system
 
 
@@ -138,7 +139,7 @@ def replace_missing_2_digit_sector_values(df):
     :param df:
     :return:
     """
-    from flowsa.flowbyfunctions import aggregator, fba_default_grouping_fields
+    from flowsa.flowbyfunctions import aggregator
 
     # check for 2 digit 0 values
     df_missing = df[(df['ActivityProducedBy'].apply(lambda x: len(x) == 2)) & (df['FlowAmount'] == 0)]
@@ -204,7 +205,7 @@ def bls_clean_allocation_fba_w_sec(df_w_sec, **kwargs):
     :return:
     """
     from flowsa.flowbyfunctions import estimate_suppressed_data, sector_disaggregation, sector_aggregation, \
-        flow_by_activity_wsec_mapped_fields, fba_mapped_default_grouping_fields
+        flow_by_activity_wsec_mapped_fields
     from flowsa.dataclean import add_missing_flow_by_fields
     from flowsa.dataclean import replace_strings_with_NoneType
