+++ flowsa/data_source_scripts/USGS_MYB_SodaAsh.py
@@ -7,7 +7,8 @@ import numpy as np
 import io
 from flowsa.common import *
 from string import digits
-from flowsa.flowbyfunctions import assign_fips_location_system , fba_default_grouping_fields, aggregator
+from flowsa.flowbyfunctions import assign_fips_location_system , aggregator
+from flowsa.common import fba_default_grouping_fields
 import math
