+++ flowsa/fbs_allocation.py
@@ -1,16 +1,19 @@
-
+import logging
 import logging as log
 import numpy as np
 import pandas as pd
 import sys
-from flowsa.common import load_source_catalog, activity_fields, US_FIPS
+
+import flowsa
+from flowsa.common import load_source_catalog, activity_fields, US_FIPS, fba_activity_fields, fbs_activity_fields, \
+    fba_mapped_default_grouping_fields, flow_by_activity_fields, fba_fill_na_dict
 from flowsa.datachecks import check_if_losing_sector_data, check_allocation_ratios, \
     check_if_location_systems_match
-from flowsa.flowbyfunctions import fba_activity_fields, fba_mapped_default_grouping_fields, collapse_activity_fields, \
-    fbs_activity_fields, sector_aggregation, sector_disaggregation
+from flowsa.flowbyfunctions import collapse_activity_fields, \
+    sector_aggregation, sector_disaggregation, allocate_by_sector, \
+    proportional_allocation_by_location_and_activity, subset_df_by_geoscale
 from flowsa.mapping import get_fba_allocation_subset
-from flowsa.dataclean import load_map_clean_fba, replace_strings_with_NoneType, \
-    replace_NoneType_with_empty_cells
+from flowsa.dataclean import replace_strings_with_NoneType, clean_df, harmonize_units
 
 # import specific functions
 from flowsa.data_source_scripts.BEA import subset_BEA_Use
@@ -295,192 +298,58 @@ def allocation_helper(df_w_sector, attr, method, v):
     return modified_fba_allocation
 
 
-def allocate_by_sector(df_w_sectors, source_name, allocation_source, allocation_method, group_cols, **kwargs):
+def load_map_clean_fba(method, attr, fba_sourcename, df_year, flowclass, geoscale_from, geoscale_to, **kwargs):
     """
-    Create an allocation ratio for df
-
-    :param df_w_sectors: df with column of sectors
-    :param source_name: the name of the FBA df being allocated
-    :param allocation_source: The name of the FBA allocation dataframe
-    :param allocation_method: currently written for 'proportional'
-    :param group_cols: columns on which to base aggregation and disaggregation
-    :return: df with FlowAmountRatio for each sector
-    """
-
-    # first determine if there is a special case with how the allocation ratios are created
-    if allocation_method == 'proportional-flagged':
-        # if the allocation method is flagged, subset sectors that are flagged/notflagged, where nonflagged sectors \
-        # have flowamountratio=1
-        if kwargs != {}:
-            if 'flowSubsetMapped' in kwargs:
-                fsm = kwargs['flowSubsetMapped']
-                flagged = fsm[fsm['disaggregate_flag'] == 1]
-                #todo: change col to be generalized, not SEctorConsumedBy specific
-                flagged_names = flagged['SectorConsumedBy'].tolist()
-
-                nonflagged = fsm[fsm['disaggregate_flag'] == 0]
-                nonflagged_names = nonflagged['SectorConsumedBy'].tolist()
-
-                # subset the original df so rows of data that run through the proportioanl allocation process are
-                # sectors included in the flagged list
-                df_w_sectors_nonflagged = df_w_sectors.loc[
-                    (df_w_sectors[fbs_activity_fields[0]].isin(nonflagged_names)) |
-                    (df_w_sectors[fbs_activity_fields[1]].isin(nonflagged_names))
-                    ].reset_index(drop=True)
-                df_w_sectors_nonflagged = df_w_sectors_nonflagged.assign(FlowAmountRatio=1)
-
-                df_w_sectors = df_w_sectors.loc[(df_w_sectors[fbs_activity_fields[0]].isin(flagged_names)) |
-                                                (df_w_sectors[fbs_activity_fields[1]].isin(flagged_names))
-                                                ].reset_index(drop=True)
-            else:
-                log.error('The proportional-flagged allocation method requires a column "disaggregate_flag" in the'
-                          'flow_subset_mapped df')
-
-    # run sector aggregation fxn to determine total flowamount for each level of sector
-    if len(df_w_sectors) == 0:
-        allocation_df = df_w_sectors_nonflagged.copy()
-    else:
-        df1 = sector_aggregation(df_w_sectors, group_cols)
-        # run sector disaggregation to capture one-to-one naics4/5/6 relationships
-        df2 = sector_disaggregation(df1, group_cols)
-
-        # if statements for method of allocation
-        # either 'proportional' or 'proportional-flagged'
-        allocation_df = []
-        if allocation_method == 'proportional' or allocation_method == 'proportional-flagged':
-            allocation_df = proportional_allocation_by_location(df2)
-        else:
-            log.error('Must create function for specified method of allocation')
-
-        if allocation_method == 'proportional-flagged':
-            # drop rows where values are not in flagged names
-            allocation_df = allocation_df.loc[(allocation_df[fbs_activity_fields[0]].isin(flagged_names)) |
-                                              (allocation_df[fbs_activity_fields[1]].isin(flagged_names))
-                                              ].reset_index(drop=True)
-            # concat the flagged and nonflagged dfs
-            allocation_df = pd.concat([allocation_df, df_w_sectors_nonflagged],
-                                      ignore_index=True).sort_values(['SectorProducedBy', 'SectorConsumedBy'])
-
-    return allocation_df
-
-
-def proportional_allocation_by_location(df):
-    """
-    Creates a proportional allocation based on all the most aggregated sectors within a location
-    Ensure that sectors are at 2 digit level - can run sector_aggregation() prior to using this function
-    :param df:
-    :param sectorcolumn:
+    Load, clean, and map a FlowByActivity df
+    :param method:
+    :param attr:
+    :param fba_sourcename:
+    :param df_year:
+    :param flowclass:
+    :param geoscale_from:
+    :param geoscale_to:
+    :param kwargs:
     :return:
     """
 
-    # tmp drop NoneType
-    df = replace_NoneType_with_empty_cells(df)
+    from flowsa.datachecks import check_if_data_exists_at_geoscale
+    from flowsa.mapping import add_sectors_to_flowbyactivity
 
-    # find the shortest length sector
+    log.info("Loading allocation flowbyactivity " + fba_sourcename + " for year " +
+             str(df_year))
+    fba = flowsa.getFlowByActivity(datasource=fba_sourcename, year=df_year, flowclass=flowclass)
+    fba = clean_df(fba, flow_by_activity_fields, fba_fill_na_dict)
+    fba = harmonize_units(fba)
 
-    denom_df = df.loc[(df['SectorProducedBy'].apply(lambda x: len(x) == 2)) |
-                      (df['SectorConsumedBy'].apply(lambda x: len(x) == 2))]
-    denom_df = denom_df.assign(Denominator=denom_df['FlowAmount'].groupby(
-        denom_df['Location']).transform('sum'))
-    denom_df_2 = denom_df[['Location', 'LocationSystem', 'Year', 'Denominator']].drop_duplicates()
-    # merge the denominator column with fba_w_sector df
-    allocation_df = df.merge(denom_df_2, how='left')
-    # calculate ratio
-    allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['FlowAmount'] / allocation_df[
-        'Denominator']
-    allocation_df = allocation_df.drop(columns=['Denominator']).reset_index()
+    # check if allocation data exists at specified geoscale to use
+    log.info("Checking if allocation data exists at the " + geoscale_from + " level")
+    check_if_data_exists_at_geoscale(fba, geoscale_from)
 
-    # add nonetypes
-    allocation_df = replace_strings_with_NoneType(allocation_df)
+    # aggregate geographically to the scale of the flowbyactivty source, if necessary
+    fba = subset_df_by_geoscale(fba, geoscale_from, geoscale_to)
 
-    return allocation_df
+    # subset based on yaml settings
+    if 'flowname_subset' in kwargs:
+            if kwargs['flowname_subset'] != 'None':
+                fba = fba.loc[fba['FlowName'].isin(kwargs['flowname_subset'])]
+    if 'compartment_subset' in kwargs:
+        if kwargs['compartment_subset'] != 'None':
+            fba = fba.loc[fba['Compartment'].isin(kwargs['compartment_subset'])]
 
+    # cleanup the fba allocation df, if necessary
+    if 'clean_fba' in kwargs:
+        log.info("Cleaning " + fba_sourcename)
+        fba = getattr(sys.modules[__name__], kwargs["clean_fba"])(fba, attr=attr)
+    # reset index
+    fba = fba.reset_index(drop=True)
 
-def proportional_allocation_by_location_and_activity(df, sectorcolumn, allocation_method):
-    """
-    Creates a proportional allocation within each aggregated sector within a location
-    :param df:
-    :param sectorcolumn:
-    :return:
-    """
+    # assign sector to allocation dataset
+    log.info("Adding sectors to " + fba_sourcename)
+    fba_wsec = add_sectors_to_flowbyactivity(fba, sectorsourcename=method['target_sector_source'])
 
-    # tmp replace NoneTypes with empty cells
-    df = replace_NoneType_with_empty_cells(df)
+    # call on fxn to further clean up/disaggregate the fba allocation data, if exists
+    if 'clean_fba_w_sec' in kwargs:
+        log.info("Further disaggregating sectors in " + fba_sourcename)
+        fba_wsec = getattr(sys.modules[__name__], kwargs['clean_fba_w_sec'])(fba_wsec, attr=attr, method=method)
 
-    # denominator summed from highest level of sector grouped by location
-    short_length = min(df[sectorcolumn].apply(lambda x: len(str(x))).unique())
-    # want to create denominator based on short_length
-    denom_df = df.loc[df[sectorcolumn].apply(lambda x: len(x) == short_length)].reset_index(drop=True)
-    grouping_cols = [e for e in ['FlowName', 'Location', 'Activity', 'ActivityConsumedBy', 'ActivityProducedBy']
-                     if e in denom_df.columns.values.tolist()]
-    denom_df.loc[:, 'Denominator'] = denom_df.groupby(grouping_cols)['HelperFlow'].transform('sum')
-
-    # list of column headers, that if exist in df, should be aggregated using the weighted avg fxn
-    possible_column_headers = ('Location', 'LocationSystem', 'Year', 'Activity', 'ActivityConsumedBy', 'ActivityProducedBy')
-    # list of column headers that do exist in the df being aggregated
-    column_headers = [e for e in possible_column_headers if e in denom_df.columns.values.tolist()]
-    merge_headers = column_headers.copy()
-    column_headers.append('Denominator')
-    # create subset of denominator values based on Locations and Activities
-    denom_df_2 = denom_df[column_headers].drop_duplicates().reset_index(drop=True)
-    # merge the denominator column with fba_w_sector df
-    allocation_df = df.merge(denom_df_2,
-                             how='left',
-                             left_on=merge_headers,
-                             right_on=merge_headers)
-    # calculate ratio
-    allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['HelperFlow'] / allocation_df['Denominator']
-    allocation_df = allocation_df.drop(columns=['Denominator']).reset_index(drop=True)
-
-    # fill empty cols with NoneType
-    allocation_df = replace_strings_with_NoneType(allocation_df)
-    # fill na values with 0
-    allocation_df['HelperFlow'] = allocation_df['HelperFlow'].fillna(0)
-
-    return allocation_df
-
-
-# def proportional_allocation_by_location_and_sector(df, sectorcolumn):
-#     """
-#     Creates a proportional allocation within each aggregated sector within a location
-#     :param df:
-#     :param sectorcolumn:
-#     :return:
-#     """
-#     from flowsa.common import load_source_catalog
-#
-#     cat = load_source_catalog()
-#     src_info = cat[pd.unique(df['SourceName'])[0]]
-#     # load source catalog to determine the level of sector aggregation associated with a crosswalk
-#     level_of_aggregation = src_info['sector_aggregation_level']
-#
-#     # denominator summed from highest level of sector grouped by location
-#     short_length = min(df[sectorcolumn].apply(lambda x: len(str(x))).unique())
-#     # want to create denominator based on short_length - 1, unless short_length = 2
-#     denom_df = df.loc[df[sectorcolumn].apply(lambda x: len(x) == short_length)]
-#     if (level_of_aggregation == 'disaggregated') & (short_length != 2):
-#         short_length = short_length - 1
-#         denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn].apply(lambda x: x[0:short_length])
-#         denom_df.loc[:, 'Denominator'] = denom_df.groupby(['Location', 'sec_tmp'])['FlowAmount'].transform('sum')
-#     else:  # short_length == 2:]
-#         denom_df.loc[:, 'Denominator'] = denom_df['FlowAmount']
-#         denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn]
-#     # if short_length == 2:
-#     #     denom_df.loc[:, 'Denominator'] = denom_df['FlowAmount']
-#     #     denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn]
-#     # else:
-#     #     short_length = short_length - 1
-#     #     denom_df.loc[:, 'sec_tmp'] = denom_df[sectorcolumn].apply(lambda x: x[0:short_length])
-#     #     denom_df.loc[:, 'Denominator'] = denom_df.groupby(['Location', 'sec_tmp'])['FlowAmount'].transform('sum')
-#
-#     denom_df_2 = denom_df[['Location', 'LocationSystem', 'Year', 'sec_tmp', 'Denominator']].drop_duplicates()
-#     # merge the denominator column with fba_w_sector df
-#     df.loc[:, 'sec_tmp'] = df[sectorcolumn].apply(lambda x: x[0:short_length])
-#     allocation_df = df.merge(denom_df_2, how='left', left_on=['Location', 'LocationSystem', 'Year', 'sec_tmp'],
-#                              right_on=['Location', 'LocationSystem', 'Year', 'sec_tmp'])
-#     # calculate ratio
-#     allocation_df.loc[:, 'FlowAmountRatio'] = allocation_df['FlowAmount'] / allocation_df[
-#         'Denominator']
-#     allocation_df = allocation_df.drop(columns=['Denominator', 'sec_tmp']).reset_index(drop=True)
-#
-#     return allocation_df
+    return fba_wsec
