+++ flowsa/flowbyactivity.py
@@ -10,7 +10,6 @@ EX: --year 2015 --source USGS_NWIS_WU
 import argparse
 from flowsa.common import *
 from esupy.processed_data_mgmt import write_df_to_file
-from flowsa.flowbyfunctions import fba_fill_na_dict
 from flowsa.dataclean import clean_df
 from flowsa.data_source_scripts.Blackhurst_IO import *
 from flowsa.data_source_scripts.BLS_QCEW import *
