+++ flowsa/mapping.py
@@ -7,8 +7,8 @@ Contains mapping functions
 import pandas as pd
 import numpy as np
 from flowsa.common import datapath, sector_source_name, activity_fields, load_source_catalog, \
-    load_sector_crosswalk, log, flow_by_activity_wsec_mapped_fields
-from flowsa.flowbyfunctions import fbs_activity_fields, fba_activity_fields, load_sector_length_crosswalk
+    load_sector_crosswalk, log, flow_by_activity_wsec_mapped_fields, fba_activity_fields
+from flowsa.flowbyfunctions import fbs_activity_fields, load_sector_length_crosswalk
 from flowsa.datachecks import replace_naics_w_naics_from_another_year
 
 def get_activitytosector_mapping(source):
